separation done
