+++ client/src/init.js
@@ -1,3 +1,4 @@
-import { getAllListener } from "./listeners/listeners.js";
+import { getAllListener, getUserListener } from "./listeners/listeners.js";
 
 getAllListener(document.getElementById('get-all'));
+getUserListener(document.getElementById('user-id'))
