+++ client/src/listeners/listeners.js
@@ -1,6 +1,9 @@
-import { getUsers } from "../handlers/handler.js";
-
+import { getUser, getUsers } from "../handlers/handler.js";
 
 export const getAllListener = (element) => {
 	element.addEventListener('click', getUsers);
+};
+
+export const getUserListener = (element) => {
+	element.addEventListener('click', getUser);
 };
