+++ index.html
@@ -8,19 +8,19 @@
 	<link href="client/styles/index.css" rel="stylesheet" />
 	</head>
 	<body>
-		<div class="page">
-			<section class="input">
-			   <h1>Welcome to user API</h1>
-			   <div class="buttons">
-			     <button id="get-all">Get all users</button>
-				 <form action="#">			
-               <input type="number" min="1" max="10" required>
-                <input type="submit" value="Get user by ID">
-               </form>
-				 </div>
+	<div class="page">
+	<section class="input">
+		<h1>Welcome to user API</h1>
+		<div class="buttons">
+			<button id="get-all">Get all users</button>
+			<form action="#">			
+            <input id="id" type="number" min="1" max="10" required>
+                <input id="user-id"  type="submit" value="Get user by ID">
+            </form>
+        </div>
 			</section>
-		  <div id="result"></div>
-	    </div>
+		<div id="result"></div>
+	</div>
 		<script type="module" src="./client/src/init.js"></script>
 	</body>
 </html>
